Rename fabric_livy_session.py to livy_result.py, remove dead tests
The file only contains dataclasses (LivySessionResult,
LivySubmissionResult) — the old name was misleading. Also removes
unit tests for the deleted LivySession class and legacy
FabricApiClient session management methods.
