+++ src/dbt/adapters/fabric/livy_result.py
