+++ tests/unit/test_fabric_api_client.py
@@ -340,145 +340,6 @@ class TestGetWarehouseId:
             client.get_warehouse_id()
 
 
-class TestLivySessionManagement:
-    @patch("dbt.adapters.fabric.fabric_api_client.requests.request")
-    def test_get_existing_livy_session_matches_by_name_and_state(self, mock_request, client):
-        client._lakehouse_id = "lh-id"
-        mock_request.return_value = _make_response(
-            200,
-            {
-                "items": [
-                    {"name": "other-session", "livyState": "idle", "id": "s1"},
-                    {"name": "test-session", "livyState": "idle", "id": "s2"},
-                ]
-            },
-        )
-
-        result = client.get_existing_livy_session()
-        assert result == "s2"
-
-    @patch("dbt.adapters.fabric.fabric_api_client.requests.request")
-    def test_get_existing_livy_session_returns_none_when_no_match(self, mock_request, client):
-        client._lakehouse_id = "lh-id"
-        mock_request.return_value = _make_response(
-            200, {"items": [{"name": "other-session", "livyState": "idle", "id": "s1"}]}
-        )
-
-        result = client.get_existing_livy_session()
-        assert result is None
-
-    @patch("dbt.adapters.fabric.fabric_api_client.requests.request")
-    def test_get_existing_livy_session_ignores_dead_sessions(self, mock_request, client):
-        client._lakehouse_id = "lh-id"
-        mock_request.return_value = _make_response(
-            200,
-            {"items": [{"name": "test-session", "livyState": "dead", "id": "s1"}]},
-        )
-
-        result = client.get_existing_livy_session()
-        assert result is None
-
-    @patch("dbt.adapters.fabric.fabric_api_client.time.sleep")
-    @patch("dbt.adapters.fabric.fabric_api_client.requests.request")
-    def test_initialize_livy_session_returns_session_id(self, mock_request, mock_sleep, client):
-        client._lakehouse_id = "lh-id"
-        mock_request.return_value = _make_response(200, {"id": "new-session-id"})
-
-        result = client.initialize_livy_session()
-        assert result == "new-session-id"
-
-    @patch("dbt.adapters.fabric.fabric_api_client.time.sleep")
-    @patch("dbt.adapters.fabric.fabric_api_client.requests.request")
-    def test_initialize_livy_session_retries_on_transient_error(
-        self, mock_request, mock_sleep, client
-    ):
-        client._lakehouse_id = "lh-id"
-        error_resp = _make_response(500, text="Server Error")
-        success_resp = _make_response(200, {"id": "new-session-id"})
-        mock_request.side_effect = [error_resp, success_resp]
-
-        result = client.initialize_livy_session()
-        assert result == "new-session-id"
-        assert mock_request.call_count == 2
-
-    @patch("dbt.adapters.fabric.fabric_api_client.time.sleep")
-    @patch("dbt.adapters.fabric.fabric_api_client.requests.request")
-    def test_initialize_livy_session_retries_on_404(self, mock_request, mock_sleep, client):
-        client._lakehouse_id = "lh-id"
-        error_resp = _make_response(404, text="Not Found")
-        success_resp = _make_response(200, {"id": "new-session-id"})
-        mock_request.side_effect = [error_resp, success_resp]
-
-        result = client.initialize_livy_session()
-        assert result == "new-session-id"
-
-    @patch("dbt.adapters.fabric.fabric_api_client.time.sleep")
-    @patch("dbt.adapters.fabric.fabric_api_client.requests.request")
-    def test_initialize_livy_session_raises_on_non_transient_error(
-        self, mock_request, mock_sleep, client
-    ):
-        client._lakehouse_id = "lh-id"
-        mock_request.return_value = _make_response(400, text="Bad Request")
-
-        with pytest.raises(FabricApiError) as exc_info:
-            client.initialize_livy_session()
-        assert exc_info.value.status_code == 400
-
-    @patch("dbt.adapters.fabric.fabric_api_client.time.sleep")
-    @patch("dbt.adapters.fabric.fabric_api_client.requests.request")
-    def test_initialize_livy_session_raises_after_max_attempts(
-        self, mock_request, mock_sleep, client
-    ):
-        client._lakehouse_id = "lh-id"
-        mock_request.return_value = _make_response(500, text="Server Error")
-
-        with pytest.raises(FabricApiError):
-            client.initialize_livy_session()
-        assert mock_request.call_count == 3
-
-    @patch("dbt.adapters.fabric.fabric_api_client.requests.request")
-    def test_get_livy_session_id_reuses_cached(self, mock_request, client):
-        client._livy_session_id = "existing-session"
-
-        result = client.get_livy_session_id()
-        assert result == "existing-session"
-        mock_request.assert_not_called()
-
-    @patch("dbt.adapters.fabric.fabric_api_client.time.sleep")
-    @patch("dbt.adapters.fabric.fabric_api_client.requests.request")
-    def test_get_livy_session_id_creates_when_no_existing(self, mock_request, mock_sleep, client):
-        client._lakehouse_id = "lh-id"
-        no_sessions = _make_response(200, {"items": []})
-        new_session = _make_response(200, {"id": "new-session-id"})
-        mock_request.side_effect = [no_sessions, new_session]
-
-        result = client.get_livy_session_id()
-        assert result == "new-session-id"
-
-    @patch("dbt.adapters.fabric.fabric_api_client.requests.request")
-    def test_delete_livy_session_clears_cached_id(self, mock_request, client):
-        client._livy_session_id = "session-to-delete"
-        client._lakehouse_id = "lh-id"
-        mock_request.return_value = _make_response(200)
-
-        client.delete_livy_session()
-
-        assert client._livy_session_id is None
-
-    @patch("dbt.adapters.fabric.fabric_api_client.requests.request")
-    def test_delete_livy_session_tolerates_404(self, mock_request, client):
-        client._livy_session_id = "session-to-delete"
-        client._lakehouse_id = "lh-id"
-        mock_request.return_value = _make_response(404, text="Not Found")
-
-        client.delete_livy_session()
-        assert client._livy_session_id is None
-
-    def test_delete_livy_session_noop_when_no_session(self, client):
-        client._livy_session_id = None
-        client.delete_livy_session()
-
-
 class TestWarehouseSnapshots:
     @patch("dbt.adapters.fabric.fabric_api_client.requests.request")
     def test_get_warehouse_snapshots_filters_by_warehouse(self, mock_request, client):
